init to base MC.c
